redesigned code
