+++ classes/average_fuel_controller.php
@@ -1,0 +1,14 @@
+<?php
+class Average_Fuel_Controller
+{
+    private $model;
+    
+    public function __construct($model) {
+        $this->model = $model;
+    }
+    
+    public function calculate($fuel, $kilo) {
+        $this->model->calculate($fuel, $kilo);
+    }
+}
+?>
